# SpringBootWebPos/batch-processing/src/main/java/com/example/batchprocessing/model/Review.java
package com.example.batchprocessing.model;

import jakarta.persistence.Entity;
import jakarta.persistence.GeneratedValue;
import jakarta.persistence.GenerationType;
import jakarta.persistence.Id;
import jakarta.persistence.Table;
import lombok.AllArgsConstructor;
import lombok.Data;

@Data
@AllArgsConstructor
@Entity
@Table(name = "reviews")
public class Review {
    
    @Id
    @GeneratedValue(strategy = GenerationType.IDENTITY)
    private Integer id;

    private Double rating;
    private String title;
    private String text;
    private String parentAsin;
    private String userId;
    private Long timestamp;

    public Review(){

    }
}